REFACTOR: Project structure. remove test files
