+++ src/app/services/pokemon.service.ts
@@ -1,12 +1,12 @@
 import { Injectable } from '@angular/core';
-import { Pokemon } from './models/pokemon.model';
+import { Pokemon } from '../models/pokemon.model';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Observable, of } from 'rxjs';
 import { catchError, map, tap } from 'rxjs/operators';
 import { MessageService } from './message.service';
-import { PokemonDetails } from './models/pokemon-details.model';
-import { POKEMON_API } from './resources';
-import { PokemonList } from './models/pokemonList.model';
+import { PokemonDetails } from '../models/pokemon-details.model';
+import { POKEMON_API } from '../resources';
+import { PokemonList } from '../models/pokemonList.model';
 
 @Injectable({
   providedIn: 'root',
